refactor: refactor textarea input
