+++ src/components/TextArea/TextArea.module.scss
@@ -1,7 +1,8 @@
 .textArea {
-  padding: 1rem;
-  font-size: 1.6rem;
   width: 50%;
   min-height: 50%;
-}
 
+  font-size: 1.6rem;
+
+  padding: 1rem;
+}
