--- src/pages/components/textArea/TextArea.types.ts
@@ -1,5 +1,0 @@
-interface ITextAreaProps {
-  value: string;
-}
-
-export default ITextAreaProps;
